Fix signature of multi_send_pinfo(...)
The second parameter was a byte, not a pointer to a struct.
(if you look at the definition of `dthread_send_delta`, you will see it is eventually assigned into an `unsigned char` array, so it can not be a pointer)

After fixing the declarations in IDA as well, it outputs code roughly equivalent to the one in this commit.