Theme_SkelRoom almost bin exact.
For some reason, our code gets too optimized, where
common subexpressions are factored out. E.g. if
xp-1 is used inside the if-then and the else body,
then in our version xp-1 is factored out of the if-else
statement and performed directly before the if-condition.

Thus the output code is shorted in our version than the
original.