Update monster.cpp
In the monster data tables, the "mType" comment has been renamed as
Seeding size. The game, as depicted by Jarulf, uses that number to
calculate how many monster type can spawn within a level (iirc the max
seeding is 4000, and each monster type takes a part of it).  I hope this
helps!