Uncomment delete to make appfat.cpp compile as C.
Note, in the future, we want to figure out how the
delete operator was overloaded and included while
still compiling the project as C.

Most likely it was overloaded in the Storm header,
and being included by appfat.cpp (which is the first
file being compiled in alphabetic order), it was
included here.

Fixes #528.