Load current heros level from save game when on difficulty screen
Fixes #566

The hero is unloaded before entering this screen from a game, hacking
the game to track the hero level would not work with single player
(planned to also have a select screen) so the only proper solution is to
load the info from the save game. Using gszHero allows us to stay synced
with the engine.