items: add BUGFIX for SpawnWitch
The logic handling the list of items for sale by Adria is broken. In
particular, the logic of SortWitch requires the list to end with an
item having item type ITYPE_NONE.

From SortWitch:
	j = 3;
	while (witchitem[j + 1]._itype != ITYPE_NONE) {
		j++;
	}

However, this puts a limit (19) on the maximum number of items being
generated for the witchitem array (of length 20). The problem is that
SpawnWitch may generate 20 items if `iCnt = 17`, as made possible by
the following statement.

	iCnt = random_(51, 8) + 10;

Since there are 3 items that are always generated (mana, full mana
and town portal), the total number of items generated by SpawnWitch
may be 20, which exceeds the limit (19), thus causing the logic of
SortWitch to go haywire and results in a buffer overflow.