objects: add BUGFIX for OperateSlainHero
The sendmsg parameter is unused in OperateSlainHero,
but it should be used in each item creation function
to sync the item creation to connected peers.

Compare the OperateSlainHero function to all other
OperateFoo functions which take a sendmsg parameter,
only OperateSlainHero ignores this parameter. All
other functions use it to sync item generation.