lighting: add BUGFIX for MakeLightTable
The i/j/k/l variables obfuscate the fact that the distance
formula used for generating data in lightblock is using the x
offset with the y component of the tile coordinate and vice-versa.

Mixing x and y this way causes light to move horizontally within
a tile when the light source is moving vertically and vice-versa.